fix warnings

+++ mindspore/core/ops/logit.cc
@@ -62,7 +62,7 @@ AbstractBasePtr LogitInfer(const abstract::AnalysisEnginePtr &, const PrimitiveP
                            const std::vector<AbstractBasePtr> &input_args) {
   MS_EXCEPTION_IF_NULL(primitive);
   auto prim_name = primitive->name();
-  const size_t input_num = 1;
+  const int64_t input_num = 1;
   (void)CheckAndConvertUtils::CheckInputArgs(input_args, kEqual, input_num, prim_name);
   auto types = LogitInferType(primitive, input_args);
   auto shapes = LogitInferShape(primitive, input_args);
